handle multiple conditions. unify rule handling
